openGL square primitive works, in 1 very specific case, under very specific circumstances. But the basis is there

+++ engine/src/general/general.cpp
@@ -74,7 +74,7 @@ Circle::Circle(Vec2 pos, float radius) {
   this->radius = radius;
 }
 
-Color::Color(int r, int g, int b) {
+Color::Color(float r, float g, float b) {
   this->r=r;
   this->g=g;
   this->b=b;
