+++ engine/src/general/general.hpp
@@ -60,11 +60,11 @@ struct Circle {
 };
 
 struct Color {
-  int r;
-  int g;
-  int b;
+  float r;
+  float g;
+  float b;
   
-  Color(int r, int g, int b);
+  Color(float r, float g, float b);
   Color();
 };
 }
